* Move the problematic test to 'mlir' and check CI status.
